Pane Decendents Fix
Objects with a super-class of Wui.pane were using the tbar and bbar of
their prototype instead of new unique properties.

Fixed combo on the hyperlink to be inline, and fixed instances where
the combo wouldn't load. Also took the z-index for the drop down out of
the CSS and put it in the JavaScript to avoid ridiculously huge
z-indexes.

Tested WUI in FF, Safari and Chrome - Still needs IE test.