Wui.Notes Fix
Wui.Notes would not be displayed when added as an ftype because they
were not being processed in Form.each(). Changed the each function to
process notes on default and have a parameter to optionally block them.

Drop shadow around windows tweaked.

Radio and Ceckbox fields fixed in non-button mode.