File Uploads via HTML 5 & bug fixes
Fixes to css, combo boxes, added the sortable option to grid columns,
added the grid to renderer arguments, fixed a bug where a grid wouldn't
resize with its containing pane, changed the default fit to allow for
better column sizing. Added the Wui version number to the object.
Allowed for different ajax options to be automatically applied when
files are being uploaded. Added Wui.forAjaxFileUpload() that turns a
regular object into a file object. Removed a "feature" in Wui.Data that
was making requests take too long. Fixed a major bug in the datepicker
that would crash the browser if the Wui got loaded on the page more than
once.