Added blank string as join method parameter
Actual change is on 331, my IDE took out whitespace which isn’t bad,
but messed up the diff. JavaScript’s join() method uses a comma by
default, but when joining the escape characters it introduced a comma
into the regex, which made the escape function attempt (unsuccessfully)
to escape commas.